backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
motor==3.6.0
pymongo==4.9.1
pydantic==2.9.2
pydantic-settings==2.6.0
pyjwt==2.9.0
passlib==1.7.4
bcrypt==4.2.0
python-multipart==0.0.12
python-dotenv==1.0.1
email-validator==2.2.0